replace the old formecolors
