+++ src/components/ViewerPage/common/SolidScene/index.tsx
@@ -1,3 +1,2 @@
-// import SolidScene from "./SolidScene"
 import ViewerScene from "./ViewerScene"
 export default ViewerScene
