+++ src/components/ViewerPage/common/SolidScene/useSolidContext.tsx
@@ -3,8 +3,7 @@ import { useMemo, useCallback } from "react"
 import Config from "components/ConfigCtx"
 import { PolyhedronCtx, OperationCtx, TransitionCtx } from "../../context"
 import { Polyhedron, Face } from "math/polyhedra"
-import { Appearance, mixColor, lighten } from "./getFormeColors"
-import getFormeColors from "./getFormeColors2"
+import getFormeColors, { Appearance, mixColor, lighten } from "./getFormeColors"
 
 // Hook that takes data from Polyhedron and Animation states and decides which to use.
 export default function useSolidContext() {
